fix flecs crash

+++ scripting/engine/Playground/ecs/components/MeshComponent.swift
@@ -1,8 +1,8 @@
 public struct MeshComponent {
     public let handle: UnsafeMutableRawPointer
-    public let meshId: UInt64
+    public let meshId: UInt16
 
-    public init(handle: UnsafeMutableRawPointer, meshId: UInt64) {
+    public init(handle: UnsafeMutableRawPointer, meshId: UInt16) {
         self.handle = handle
         self.meshId = meshId
     }
